Handle namespace explicitly
The notebook references include several explicit references to the
namespace name. In order to handle that correctly, this commit adds a new
NAMESPACE parameter to the template and updates the create_notebook.sh
command such that `-n` is used to specify a namespace and `-t` now sets the
test run name.

The namespace defaults to the current namespace (retrieved from the output
of `oc config view --minify`).